+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 - Change GE Proton version from GE-Proton7-37 to GE-Proton7-55.
 - Change GE Wine version from GE-Proton7-31 to GE-Proton7-43.
 - Change the screenshot program from Shutter to Spectacle.
+- Add support for the Steam Deck controller.
 
 ## 3.2.1 - The SteamOS 3.4 Update